backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
xlsxwriter
jinja2
python-multipart
matplotlib
python-docx
google-genai